javatari/requirements.txt
audioread==3.0.1
boto3==1.12.36
botocore==1.15.36
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
click==7.1.1
decorator==5.1.1
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.2
Flask-Session==0.3.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==1.4.0
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==1.1.1
msgpack==1.0.8
nose==1.3.7
numba==0.59.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.0
pooch==1.8.1
progress==1.6
protobuf==3.6.1
pycparser==2.22
python-dateutil==2.8.1
python-vlc==3.0.7110
requests==2.23.0
s3transfer==0.3.3
scikit-learn==1.4.2
scipy==1.13.0
setuptools==68.2.2
six==1.14.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==1.25.8
virtualenv==16.7.7
Werkzeug==0.16.0
wheel==0.41.2
